+++ GameBoard.hpp
@@ -8,6 +8,7 @@ class GameBoard{
 		std::vector<Player*> players;
 	public:
 		GameBoard(int n);
+		~GameBoard();
 		void printGameStatistics();
 		std::vector<Player *>& getPlayers();
 		void printPlayers();
