+++ Stronghold.cpp
